test: Amount 금액 변환 테스트시 반환 타입인 Double로 검증

+++ src/test/kotlin/lotto/domain/AmountSpec.kt
@@ -28,7 +28,7 @@ class AmountSpec : FunSpec({
 
         val result = amount.toDouble()
 
-        result shouldBe 3000
+        result shouldBe 3000.toDouble()
     }
 
     context("총합 계산") {
